re-add hello.rs
